requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.1
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-filter==24.1
django-ipware==6.0.3
django-session-timeout==0.1.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
drf-writable-nested==0.7.0
filelock==3.13.1
filetype==1.2.0
idna==3.6
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
PyMySQL==1.1.0
python-decouple==3.8
python-ipware==2.0.1
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.1
tzdata==2023.4
urllib3==2.1.0
